Refactor Inference modules
